conan_lower_case/requirements.txt
conan==1.0.0-beta3